Replace encryption key from `.travis.yml`
Use encryption key generated based on an access token created from the
profile of the machine user¹ `h5bp-bot` (https://github.com/h5bp-bot).

This change is made in order to allow all users that have access to the
`h5bp-bot` account to delete or edit the token based on which the
encryption key is generated.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ https://developer.github.com/guides/managing-deploy-keys/#machine-users